src/Builder4/MainWindow.bjs src/Builder4/MainWindow.vala

+++ src/Builder4/MainWindow.vala
@@ -2004,6 +2004,11 @@ public class Xcls_MainWindow : Object
 
             {
                  Resources.singleton().updateProgress.connect((pos) => {
+                    if (pos < 1) {
+                        this.el.hide();
+                        return;
+                    }
+                    this.el.show();
                     var cid = this.el.get_context_id("resources");
                     this.el.remove_all(cid);
                     this.el.push(cid,"Fetching Resource : %d/6".print(pos));
